refactor(cubesql): Extract redshift UDFs to separe module
