+++ rust/cubesql/cubesql/src/compile/engine/udf/mod.rs
@@ -1,6 +1,8 @@
 mod common;
 mod pg_catalog;
+mod redshift;
 mod utils;
 
 pub use common::*;
 pub use pg_catalog::*;
+pub use redshift::*;
